Split transaction handling code out into base class
